pyproject.toml
[tool.poetry]
name = "tasktracker"
version = "0.1.0"
description = "My pet-project for learning of FastAPI + alembic + sqlachemy + pydantic + poetry"
authors = ["ImperatorNeron <katebarspk@gmail.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.1"
pydantic = {extras = ["email"], version = "^2.8.2"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.31"}
alembic = "^1.13.2"
poetry = "^1.8.3"
fastapi-users = {extras = ["sqlalchemy"], version = "^13.0.0"}
asyncpg = "^0.29.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
pydantic-settings = "^2.3.4"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
